# Sparsity-aware clipper achieves 6x speed-up on bunny example