--- a/examples/python/ex4_bunny.ipynb
+++ b/examples/python/ex4_bunny.ipynb
@@ -9,7 +9,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 15,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -22,7 +22,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 16,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -96,7 +96,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 17,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -118,11 +118,21 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 31,
    "metadata": {},
-   "outputs": [],
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "Affinity matrix creation took 0.066 seconds\n",
+      "CLIPPER selected 41 inliers from 1000 putative associations (precision 1.00, recall 0.82) in 0.085 s\n"
+     ]
+    }
+   ],
    "source": [
     "params = clipper.Params()\n",
+    "params.beta = 0.25 # this was the value in the original version of the repo\n",
     "iparams = clipper.invariants.EuclideanDistanceParams()\n",
     "iparams.sigma = 0.015\n",
     "iparams.epsilon = 0.02\n",
@@ -136,6 +146,7 @@
     "soln = clipper.find_dense_cluster(M, C, params)\n",
     "t1 = time.time()\n",
     "Ain = clipper.select_inlier_associations(soln, A)\n",
+    "dense_duration = t1 - t0\n",
     "\n",
     "p = np.isin(Ain, Agt)[:,0].sum() / Ain.shape[0]\n",
     "r = np.isin(Ain, Agt)[:,0].sum() / Agt.shape[0]\n",
@@ -145,9 +156,20 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 32,
    "metadata": {},
-   "outputs": [],
+   "outputs": [
+    {
+     "data": {
+      "text/plain": [
+       "PointCloud with 1250 points."
+      ]
+     },
+     "execution_count": 32,
+     "metadata": {},
+     "output_type": "execute_result"
+    }
+   ],
    "source": [
     "model = o3d.geometry.PointCloud()\n",
     "model.points = o3d.utility.Vector3dVector(D1.T)\n",
@@ -162,7 +184,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 33,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -172,7 +194,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 34,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -185,16 +207,27 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 35,
    "metadata": {},
-   "outputs": [],
+   "outputs": [
+    {
+     "data": {
+      "text/plain": [
+       "(0.0063437045150311604, 0.0011663986322407618)"
+      ]
+     },
+     "execution_count": 35,
+     "metadata": {},
+     "output_type": "execute_result"
+    }
+   ],
    "source": [
     "get_err(T_21, That_21)"
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 36,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -210,7 +243,95 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 37,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "draw_registration_result(model, data, That_21)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "### Exploiting Sparsity in Affinity matrix\n",
+    "\n",
+    "we run the above with sparsity-aware version of clipper"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 38,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "Affinity matrix creation took 0.063 seconds\n",
+      "sparse-aware CLIPPER selected 40 inliers from 1000 putative associations (precision 1.00, recall 0.80) in 0.012 s\n",
+      "Speed-up: 6.964860907759883\n"
+     ]
+    }
+   ],
+   "source": [
+    "params = clipper.Params()\n",
+    "iparams = clipper.invariants.EuclideanDistanceParams()\n",
+    "iparams.sigma = 0.015\n",
+    "iparams.epsilon = 0.02\n",
+    "invariant = clipper.invariants.EuclideanDistance(iparams)\n",
+    "\n",
+    "t0 = time.time()\n",
+    "M, C = clipper.score_sparse_pairwise_consistency(invariant, D1, D2, A)\n",
+    "t1 = time.time()\n",
+    "print(f\"Affinity matrix creation took {t1-t0:.3f} seconds\")\n",
+    "t0 = time.time()\n",
+    "soln = clipper.find_dense_cluster_of_sparse_graph(M, C, params)\n",
+    "t1 = time.time()\n",
+    "Ain = clipper.select_inlier_associations(soln, A)\n",
+    "sparse_duration = t1 - t0\n",
+    "\n",
+    "p = np.isin(Ain, Agt)[:,0].sum() / Ain.shape[0]\n",
+    "r = np.isin(Ain, Agt)[:,0].sum() / Agt.shape[0]\n",
+    "print(f\"sparse-aware CLIPPER selected {Ain.shape[0]} inliers from {A.shape[0]} \"\n",
+    "      f\"putative associations (precision {p:.2f}, recall {r:.2f}) in {t1-t0:.3f} s\")\n",
+    "\n",
+    "print(f\"Speed-up: {dense_duration / sparse_duration}\")"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 39,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "p2p = o3d.pipelines.registration.TransformationEstimationPointToPoint()\n",
+    "That_21 = p2p.compute_transformation(model, data, o3d.utility.Vector2iVector(Ain))"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 40,
+   "metadata": {},
+   "outputs": [
+    {
+     "data": {
+      "text/plain": [
+       "(0.006657570406907475, 0.0011222630927085323)"
+      ]
+     },
+     "execution_count": 40,
+     "metadata": {},
+     "output_type": "execute_result"
+    }
+   ],
+   "source": [
+    "get_err(T_21, That_21)"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 41,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -379,7 +500,7 @@
    "name": "python",
    "nbconvert_exporter": "python",
    "pygments_lexer": "ipython3",
-   "version": "3.6.9"
+   "version": "3.6.13"
   }
  },
  "nbformat": 4,