… include/clipper/affinity.h