Hello @plusk01,

Enjoyed the paper and the examples in this repo. Nice work!

I took the time to write and contribute a spare-aware version of clipper (sparsity here is with regard to the consistency score matrices M and C). This version achieves ~6x speed-up factor on the bunny (python notebook) example and almost 2.5x on the LargePointCloud example in the tests (see below for exact numbers). 

For Bunny python example:

> Affinity matrix creation took 0.062 seconds
> CLIPPER selected 41 inliers from 1000 putative associations (precision 1.00, recall 0.82) in 0.092 s

with sparse-aware clipper:

> Affinity matrix creation took 0.063 seconds
> sparse-aware CLIPPER selected 40 inliers from 1000 putative associations (precision 1.00, recall 0.80) in 0.014 s
> Speed-up: 6.780588566204909


For C++ Tests:

<img src="https://user-images.githubusercontent.com/1083829/157089342-9741b8c1-c179-41f1-8112-d960b8a22716.png" width="350">




Let me know your thoughts and if you have any questions.

Best,
Abdullah