@plusk01 2022-03-07
thanks @ash-aldujaili, this looks very nice - I will take a closer look and try on some other examples over the next few days