@plusk01 2022-03-22
I had a chance to look closely at your changes and benchmark them today - thanks again for your contribution!

Using the new benchmark suite (#4), I find the following results, confirming the ~6x speedup

### Dense and using mat-vec multiplies (0df870aee12b7f09e8460f5ca99a1f670bdbf2b9)

```
Benchmarking over 20 trials


Benchmarking ρ = 90% [██████████████████████████████████████████████████] 100% [01m:18s]               

+-------+---------+---------------+-------------------+---------------+------------+
| ρ [%] | # assoc | affinity [ms] | dense clique [ms] | precision [%] | recall [%] |
+-------+---------+---------------+-------------------+---------------+------------+
| 0     | 64      |  0.29  ±  0.3 |      0.03  ±  0.0 | 100           | 89         |
| 0     | 256     |  0.87  ±  0.3 |      0.33  ±  0.0 | 100           | 89         |
| 0     | 512     |  2.92  ±  0.3 |      1.74  ±  0.0 | 100           | 89         |
| 0     | 1024    | 22.67  ±  1.2 |     11.35  ±  0.0 | 100           | 89         |
| 0     | 2048    | 92.94  ±  2.2 |     74.24  ±  4.2 | 100           | 89         |
+-------+---------+---------------+-------------------+---------------+------------+
| 20    | 64      |  0.23  ±  0.1 |      0.07  ±  0.0 | 100           | 89         |
| 20    | 256     |  0.85  ±  0.1 |      1.16  ±  0.1 | 100           | 89         |
| 20    | 512     |  3.12  ±  0.3 |      8.82  ±  1.3 | 100           | 89         |
| 20    | 1024    | 22.80  ±  0.7 |     50.64  ±  6.9 | 99            | 89         |
| 20    | 2048    | 93.32  ±  4.8 |    361.77  ± 42.9 | 99            | 89         |
+-------+---------+---------------+-------------------+---------------+------------+
| 40    | 64      |  0.22  ±  0.0 |      0.10  ±  0.0 | 100           | 90         |
| 40    | 256     |  1.18  ±  1.0 |      1.45  ±  0.2 | 100           | 89         |
| 40    | 512     |  3.49  ±  0.3 |     10.39  ±  1.3 | 99            | 89         |
| 40    | 1024    | 24.55  ±  4.3 |     63.26  ±  7.8 | 100           | 89         |
| 40    | 2048    | 93.71  ±  2.8 |    435.69  ± 58.0 | 99            | 89         |
+-------+---------+---------------+-------------------+---------------+------------+
| 80    | 64      |  2.75  ±  4.1 |      0.14  ±  0.0 | 100           | 91         |
| 80    | 256     |  0.85  ±  0.0 |      1.65  ±  0.2 | 99            | 90         |
| 80    | 512     |  4.20  ±  1.7 |     13.10  ±  2.1 | 99            | 90         |
| 80    | 1024    | 23.09  ±  1.5 |     73.01  ± 10.1 | 99            | 90         |
| 80    | 2048    | 92.59  ±  2.8 |   492.38  ± 102.3 | 99            | 89         |
+-------+---------+---------------+-------------------+---------------+------------+
| 90    | 64      |  3.00  ±  5.8 |      0.23  ±  0.0 | 100           | 95         |
| 90    | 256     |  1.80  ±  2.8 |      1.95  ±  0.2 | 99            | 90         |
| 90    | 512     |  4.30  ±  1.3 |     14.59  ±  1.7 | 99            | 91         |
| 90    | 1024    | 23.52  ±  2.9 |     76.52  ±  8.2 | 99            | 90         |
| 90    | 2048    | 92.05  ±  3.3 |   526.04  ± 102.0 | 99            | 90         |
+-------+---------+---------------+-------------------+---------------+------------+
```

### Sparse and avoiding mat-vec multiplies

```
Benchmarking over 20 trials


Benchmarking ρ = 90% [██████████████████████████████████████████████████] 100% [00m:56s]               

+-------+---------+----------------+-------------------+---------------+------------+
| ρ [%] | # assoc |  affinity [ms] | dense clique [ms] | precision [%] | recall [%] |
+-------+---------+----------------+-------------------+---------------+------------+
| 0     | 64      |   0.52  ±  0.9 |      0.07  ±  0.0 | 100           | 89         |
| 0     | 256     |   1.41  ±  0.4 |      0.92  ±  0.2 | 100           | 89         |
| 0     | 512     |   5.30  ±  0.3 |      3.62  ±  0.8 | 100           | 89         |
| 0     | 1024    |  42.30  ±  1.7 |      9.63  ±  1.5 | 100           | 89         |
| 0     | 2048    | 174.11  ±  9.8 |     47.51  ±  9.4 | 100           | 89         |
+-------+---------+----------------+-------------------+---------------+------------+
| 20    | 64      |   0.25  ±  0.0 |      0.24  ±  0.0 | 100           | 89         |
| 20    | 256     |   1.18  ±  0.1 |      3.89  ±  0.8 | 100           | 89         |
| 20    | 512     |   5.10  ±  0.3 |     15.27  ±  2.5 | 99            | 89         |
| 20    | 1024    |  23.18  ±  2.5 |     54.81  ± 11.1 | 99            | 89         |
| 20    | 2048    | 133.73  ± 12.5 |    345.16  ± 77.3 | 99            | 89         |
+-------+---------+----------------+-------------------+---------------+------------+
| 40    | 64      |   5.34  ±  5.6 |      0.25  ±  0.0 | 100           | 89         |
| 40    | 256     |   1.04  ±  0.1 |      3.36  ±  0.7 | 99            | 89         |
| 40    | 512     |   4.22  ±  0.4 |     12.42  ±  1.5 | 100           | 89         |
| 40    | 1024    |  21.38  ±  3.6 |     48.93  ± 11.1 | 99            | 89         |
| 40    | 2048    |  92.75  ±  4.5 |    264.75  ± 55.5 | 99            | 89         |
+-------+---------+----------------+-------------------+---------------+------------+
| 80    | 64      |   0.37  ±  0.6 |      0.16  ±  0.0 | 99            | 91         |
| 80    | 256     |   0.86  ±  0.1 |      1.71  ±  0.2 | 100           | 90         |
| 80    | 512     |   3.46  ±  0.3 |      7.44  ±  1.7 | 99            | 89         |
| 80    | 1024    |  15.76  ±  0.4 |     23.73  ±  3.6 | 99            | 89         |
| 80    | 2048    |  63.44  ±  4.6 |    113.56  ± 28.4 | 99            | 89         |
+-------+---------+----------------+-------------------+---------------+------------+
| 90    | 64      |   0.23  ±  0.0 |      0.17  ±  0.0 | 99            | 91         |
| 90    | 256     |   0.86  ±  0.1 |      1.64  ±  0.2 | 99            | 90         |
| 90    | 512     |   3.22  ±  0.4 |      6.25  ±  1.0 | 99            | 90         |
| 90    | 1024    |  15.01  ±  0.5 |     21.29  ±  2.9 | 99            | 90         |
| 90    | 2048    |  61.48  ±  3.3 |     93.10  ± 13.6 | 99            | 90         |
+-------+---------+----------------+-------------------+---------------+------------+
```